agent/pyproject.toml
[project]
name = "ora2pg-conv-w-strands"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.38.18",
    "oracledb>=3.1.1",
    "strands-agents>=0.1.2",
    "strands-agents-tools>=0.1.1",
]
